# packages/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LinguaQuiz API",
    "description": "Language learning quiz backend with automated spaced repetition",
    "version": "4.0.0"
  },
  "paths": {
...
    "/api/word-lists": {
      "get": {
        "tags": ["Vocabulary"],
        "summary": "Get Word Lists",
        "operationId": "get_word_lists_api_word_lists_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WordListResponse"
                  },
                  "type": "array",
                  "title": "Response Get Word Lists Api Word Lists Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/translations": {
      "get": {
        "tags": ["Vocabulary"],
        "summary": "Get Translations",
        "operationId": "get_translations_api_translations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "list_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "List Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VocabularyItemResponse"
                  },
                  "title": "Response Get Translations Api Translations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/vocabulary/{item_id}": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get Vocabulary Item",
        "operationId": "get_vocabulary_item_api_admin_vocabulary__item_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VocabularyItemDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Admin"],
        "summary": "Update Vocabulary Item",
        "operationId": "update_vocabulary_item_api_admin_vocabulary__item_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VocabularyItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Admin"],
        "summary": "Delete Vocabulary Item",
        "operationId": "delete_vocabulary_item_api_admin_vocabulary__item_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/vocabulary": {
      "post": {
        "tags": ["Admin"],
        "summary": "Create Vocabulary Item",
        "operationId": "create_vocabulary_item_api_admin_vocabulary_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VocabularyItemCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Admin"],
        "summary": "List Vocabulary",
        "operationId": "list_vocabulary_api_admin_vocabulary_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "list_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "List Name"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VocabularyItemDetailResponse"
                  },
                  "title": "Response List Vocabulary Api Admin Vocabulary Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
      "VocabularyItemCreate": {
        "properties": {
          "sourceText": {
            "type": "string",
            "title": "Sourcetext"
          },
          "sourceLanguage": {
            "type": "string",
            "title": "Sourcelanguage"
          },
          "targetText": {
            "type": "string",
            "title": "Targettext"
          },
          "targetLanguage": {
            "type": "string",
            "title": "Targetlanguage"
          },
          "listName": {
            "type": "string",
            "title": "Listname"
          },
          "difficultyLevel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficultylevel"
          },
          "sourceUsageExample": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceusageexample"
          },
          "targetUsageExample": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetusageexample"
          }
        },
        "type": "object",
        "required": ["sourceText", "sourceLanguage", "targetText", "targetLanguage", "listName"],
        "title": "VocabularyItemCreate"
      },
      "VocabularyItemDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "sourceText": {
            "type": "string",
            "title": "Sourcetext"
          },
          "sourceLanguage": {
            "type": "string",
            "title": "Sourcelanguage"
          },
          "targetText": {
            "type": "string",
            "title": "Targettext"
          },
          "targetLanguage": {
            "type": "string",
            "title": "Targetlanguage"
          },
          "listName": {
            "type": "string",
            "title": "Listname"
          },
          "difficultyLevel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficultylevel"
          },
          "sourceUsageExample": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceusageexample"
          },
          "targetUsageExample": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetusageexample"
          },
          "isActive": {
            "type": "boolean",
            "title": "Isactive"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "sourceText",
          "sourceLanguage",
          "targetText",
          "targetLanguage",
          "listName",
          "difficultyLevel",
          "sourceUsageExample",
          "targetUsageExample",
          "isActive",
          "createdAt",
          "updatedAt"
        ],
        "title": "VocabularyItemDetailResponse"
      },
      "VocabularyItemResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "sourceText": {
            "type": "string",
            "title": "Sourcetext"
          },
          "sourceLanguage": {
            "type": "string",
            "title": "Sourcelanguage"
          },
          "targetText": {
            "type": "string",
            "title": "Targettext"
          },
          "targetLanguage": {
            "type": "string",
            "title": "Targetlanguage"
          },
          "listName": {
            "type": "string",
            "title": "Listname"
          },
          "sourceUsageExample": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceusageexample"
          },
          "targetUsageExample": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetusageexample"
          }
        },
        "type": "object",
        "required": [
          "id",
          "sourceText",
          "sourceLanguage",
          "targetText",
          "targetLanguage",
          "listName",
          "sourceUsageExample",
          "targetUsageExample"
        ],
        "title": "VocabularyItemResponse"
      },
      "VocabularyItemUpdate": {
        "properties": {
          "sourceText": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcetext"
          },
          "targetText": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targettext"
          },
          "sourceUsageExample": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceusageexample"
          },
          "targetUsageExample": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetusageexample"
          },
          "isActive": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isactive"
          }
        },
        "type": "object",
        "title": "VocabularyItemUpdate"
      },
      "WordListResponse": {
        "properties": {
          "listName": {
            "type": "string",
            "title": "Listname"
          },
          "wordCount": {
            "type": "integer",
            "title": "Wordcount"
          }
        },
        "type": "object",
        "required": ["listName", "wordCount"],
        "title": "WordListResponse"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}